+++ lib/karafka/connection/proxy.rb
@@ -10,20 +10,13 @@ module Karafka
     # do still want to be able to alter some functionalities. This wrapper helps us do it when
     # it would be needed
     class Proxy < SimpleDelegator
-      # Timeout on the watermark query
-      WATERMARK_REQUEST_TIMEOUT = 5_000
+      # Errors on which we want to retry
+      RETRYABLE_ERRORS = %i[
+        all_brokers_down
+        timed_out
+      ].freeze
 
-      # Timeout on the TPL request query
-      TPL_REQUEST_TIMEOUT = 5_000
-
-      # How many attempts we want to take for something that would end up with all_brokers_down
-      BROKERS_DOWN_MAX_ATTEMPTS = 3
-
-      # How long should we wait in between all_brokers_down
-      BROKERS_DOWN_BACKOFF_TIME = 1
-
-      private_constant :WATERMARK_REQUEST_TIMEOUT, :BROKERS_DOWN_MAX_ATTEMPTS,
-                       :BROKERS_DOWN_BACKOFF_TIME, :TPL_REQUEST_TIMEOUT
+      private_constant :RETRYABLE_ERRORS
 
       attr_accessor :wrapped
 
@@ -36,6 +29,7 @@ module Karafka
         # wrap an already wrapped object with another proxy level. Simplifies passing consumers
         # and makes it safe to wrap without type checking
         @wrapped = obj.is_a?(self.class) ? obj.wrapped : obj
+        @config = ::Karafka::App.config.internal.connection.proxy
       end
 
       # Proxies the `#query_watermark_offsets` with extra recovery from timeout problems.
@@ -46,12 +40,14 @@ module Karafka
       # @param partition [Partition]
       # @return [Array<Integer, Integer>] watermark offsets
       def query_watermark_offsets(topic, partition)
-        with_brokers_down_retry do
-          @wrapped.query_watermark_offsets(
-            topic,
-            partition,
-            WATERMARK_REQUEST_TIMEOUT
-          )
+        l_config = @config.query_watermark_offsets
+
+        with_broker_errors_retry(
+          # required to be in seconds, not ms
+          wait_time: l_config.wait_time / 1_000.to_f,
+          max_attempts: l_config.max_attempts
+        ) do
+          @wrapped.query_watermark_offsets(topic, partition, l_config.timeout)
         end
       end
 
@@ -61,8 +57,14 @@ module Karafka
       # @param tpl [Rdkafka::Consumer::TopicPartitionList] tpl to get time offsets
       # @return [Rdkafka::Consumer::TopicPartitionList] tpl with time offsets
       def offsets_for_times(tpl)
-        with_brokers_down_retry do
-          @wrapped.offsets_for_times(tpl, TPL_REQUEST_TIMEOUT)
+        l_config = @config.offsets_for_times
+
+        with_broker_errors_retry(
+          # required to be in seconds, not ms
+          wait_time: l_config.wait_time / 1_000.to_f,
+          max_attempts: l_config.max_attempts
+        ) do
+          @wrapped.offsets_for_times(tpl, l_config.timeout)
         end
       end
 
@@ -71,16 +73,20 @@ module Karafka
       # Runs expected block of code with few retries on all_brokers_down
       # librdkafka can return `all_brokers_down` for scenarios when broker is overloaded or not
       # reachable due to latency.
-      def with_brokers_down_retry
+      # @param max_attempts [Integer] how many attempts (not retries) should we take before failing
+      #   completely.
+      # @param wait_time [Integer, Float] how many seconds should we wait. It uses `#sleep` of Ruby
+      #   so it needs time in seconds.
+      def with_broker_errors_retry(max_attempts:, wait_time: 1)
         attempt ||= 0
         attempt += 1
 
         yield
       rescue Rdkafka::RdkafkaError => e
-        raise if e.code != :all_brokers_down
+        raise unless RETRYABLE_ERRORS.include?(e.code)
 
-        if attempt <= BROKERS_DOWN_MAX_ATTEMPTS
-          sleep(BROKERS_DOWN_BACKOFF_TIME)
+        if attempt <= max_attempts
+          sleep(wait_time)
 
           retry
         end